ok, fixed south
I think issue is my conversion between vector29, vector31, and row and column. It's way too confusing to keep the mapping in my head of "Vector29 is the list of rows, but you use him as column? no wait, row?". Better to make conversion functions based on the successful south movement.